requirements.txt
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
tzdata==2022.2
